Updated the HeroesComponent.
Deleted the HEROES (constant) import, and imported the HeroService instead. Replaced the definition of the heroes property with a declaration as an array of Hero type.

Injected the HeroService into the HeroesComponent class by adding in a private heroService parameter of type HeroService into the constructor.

Added a getHeroes() method to retrieve the heroes form the heroesService.