Updated dashboard component. Imported the OnInit symbol from '@angular/core', imported the Hero type interface, and imported HeroService.
Added a constructer with a private heroService of type HeroService for injection.

Added ngOnInt() method to call the getHeroes() method when the dashboard component is created.

Added the getHeroes() method to subscribe and retrieve data from the asynchronous heroService.getHeroes() method call and add slice of the top 5 heroes into the empty heroes array of type Hero.