Updated the getHeroes() method to tap into the flow of the observable values and send a message, using the log() method, to the message area at the bottom of the page.
The HeroService methods taps into the flow of observable values and send a message, using the log() method, to the message area at the bottom of the page.

The RxJS tap() operator enables this ability by looking at the observable values, doing something with those values, and passing them along.