Updated the getHero() method to  to construct a request URL with the desired hero's id.
Most web AP ls support a get by id request in the form : baseURL/ : id . Here, the base URL is the heroesURL defined in the Heroes and HTTP section in api/heroes and id is the number of the hero that you want to retrieve.

The server should respond with a single hero rather than an array of heroes. getHero( ) returns an Obse rvable<Hero>, which is an observable of Hero objects rather than an observable of Hero arrays.